requirements.txt
asyncmy==0.2.5
sqlalchemy==1.4.35
tornado==6.1
marshmallow==3.15.0
python-json-logger==2.0.2
aioredis==1.3.1
python-dateutil==2.8.2
pytest==7.1.1
pytest-asyncio==0.18.3
pytest-mock==3.7.0
